fix(build): align ConfigurationModels types with preferences, fix icon reference

+++ app/src/main/java/dev/aurakai/auraframefx/domains/aura/ui/screens/manual/LaunchMatrixManualScreen.kt
@@ -119,7 +119,7 @@ fun LaunchMatrixManualScreen(
                     onValueChange = {
                         viewModel.updateLauncherConfig(
                             context,
-                            config.copy(iconSize = (it / 100f).toInt())
+                            config.copy(iconSize = it / 100f)
                         )
                     }
                 )
